// src/client/api/internal/schemas/listNotesItem.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * notesnet
 * OpenAPI spec version: 0.1.0
 */

export interface ListNotesItem {
  attributes: unknown;
  content: unknown;
  content_type: string;
  inserted_at: string;
  note_author_persona_id: string;
  note_id: string;
  revision_author_persona_id: string;
  revision_id: string;
  scope_id: string;
  /** @nullable */
  started_at?: string | null;
  /** @nullable */
  summary?: string | null;
  text_for_search: string;
  written_at: string;
}